Split owned canvas preview responsibilities
