+++ packages/builder-preview/src/owned-canvas.ts
@@ -1,3 +1,8 @@
+import { mirrorPopupGeometry as updatePopupGeometry } from './owned-canvas/geometry';
+import { isEnabled, selectorAttributeFilter } from './owned-canvas/settings';
+import { attachCanvasStyles, copyPopupStyles } from './owned-canvas/styles';
+import { getCanvasTransformSignature } from './owned-canvas/transforms';
+
 const ownedCanvas = window.pumBuilderOwnedCanvas;
 
 if ( ownedCanvas?.popup_id ) {
@@ -32,27 +37,6 @@ if ( ownedCanvas?.popup_id ) {
 			originalValue: string;
 		}
 	>();
-	// Mirrors the responsive presets in the site stylesheet, using viewport
-	// units because a builder canvas may have a narrower containing block.
-	const responsiveWidths: Record< string, string > = {
-		nano: '10vw',
-		micro: '20vw',
-		tiny: '30vw',
-		small: '40vw',
-		medium: '60vw',
-		normal: '70vw',
-		large: '80vw',
-		xlarge: '95vw',
-	};
-
-	const isEnabled = ( value: boolean | string ): boolean =>
-		true === value || '1' === value;
-
-	const numericOffset = ( value: string ): number => {
-		const offset = parseFloat( value );
-
-		return Number.isNaN( offset ) ? 0 : offset;
-	};
 
 	const rememberOwnedCanvasAttribute = (
 		element: Element,
@@ -155,138 +139,6 @@ if ( ownedCanvas?.popup_id ) {
 				addRootClass( targetDocument.documentElement, className );
 				addRootClass( targetDocument.body, className );
 			} );
-	};
-
-	const selectorAttributeFilter = (
-		...selectors: Array< string | undefined >
-	): string[] => {
-		const attributes = [ 'class', 'id' ];
-		const pattern = /\[\s*([^\s~|^$*=\]]+)/g;
-
-		selectors.forEach( ( selector ) => {
-			if ( ! selector ) {
-				return;
-			}
-
-			let match: RegExpExecArray | null;
-
-			while ( ( match = pattern.exec( selector ) ) ) {
-				attributes.push( match[ 1 ] );
-			}
-		} );
-
-		return Array.from( new Set( attributes ) );
-	};
-
-	const viewportRelativeLength = (
-		value: string,
-		viewportUnit = 'vw'
-	): string => {
-		const percentage = value.trim().match( /^(\d+(?:\.\d+)?)%$/ );
-
-		return percentage ? `${ percentage[ 1 ] }${ viewportUnit }` : value;
-	};
-
-	const responsiveWidth = ( size: string, viewportWidth: number ): string => {
-		if ( viewportWidth < 1024 ) {
-			return '95vw';
-		}
-
-		return responsiveWidths[ size ] || '95vw';
-	};
-
-	const copyPopupStyles = ( targetDocument: Document ): void => {
-		( display.style_selectors || [] ).forEach(
-			( selector, selectorIndex ) => {
-				document
-					.querySelectorAll( selector )
-					.forEach( ( source, sourceIndex ) => {
-						const sourceId = source.getAttribute( 'id' );
-						const copyId = sourceId
-							? `pum-builder-copy-${ sourceId }`
-							: `pum-builder-copy-anonymous-${ selectorIndex }-${ sourceIndex }`;
-
-						if ( targetDocument.getElementById( copyId ) ) {
-							return;
-						}
-
-						const copy = source.cloneNode( true ) as HTMLElement;
-
-						copy.id = copyId;
-
-						if ( 'LINK' === copy.tagName ) {
-							copy.addEventListener( 'load', adoptCanvas );
-						}
-						targetDocument.head.append( copy );
-						ownedStyleElements.push( copy );
-					} );
-			}
-		);
-	};
-
-	const attachCanvasStyles = ( targetDocument: Document ): void => {
-		if ( targetDocument.getElementById( 'pum-builder-owned-canvas' ) ) {
-			return;
-		}
-
-		const style = targetDocument.createElement( 'style' );
-		style.id = 'pum-builder-owned-canvas';
-		style.textContent = `
-			html.pum-builder-owned-canvas-root > body {
-				background: transparent !important;
-			}
-			html.pum-builder-owned-canvas-root.pum-overlay-disabled,
-			html.pum-builder-owned-canvas-root > body.pum-overlay-disabled {
-				background: transparent !important;
-			}
-			html.pum-builder-owned-canvas {
-				display: block !important;
-				min-height: 100%;
-				opacity: 1 !important;
-				overflow: auto !important;
-				visibility: visible !important;
-			}
-			html.pum-builder-owned-canvas.pum-overlay-disabled {
-				background: transparent !important;
-			}
-			html.pum-builder-owned-canvas > body.pum-builder-canvas-area {
-				display: block !important;
-				height: auto;
-				margin: 0 !important;
-				min-height: 0 !important;
-				opacity: 1 !important;
-				position: absolute !important;
-				visibility: visible !important;
-			}
-			.pum-builder-canvas-title,
-			.pum-builder-canvas-close {
-				pointer-events: none !important;
-			}
-			.pum-builder-canvas-close-anchor {
-				display: none !important;
-			}
-			.pum-builder-canvas-area,
-			.pum-builder-canvas-area *,
-			.pum-builder-canvas-area *::before,
-			.pum-builder-canvas-area *::after {
-				box-sizing: border-box;
-			}
-			.pum-builder-canvas-area.pum-content:focus {
-				outline: none;
-			}
-			.pum-builder-canvas-area.pum-content > :nth-child(1 of :not(.pum-builder-canvas-title, .pum-builder-canvas-close-anchor, .pum-builder-canvas-close)) {
-				margin-top: 0;
-			}
-			.pum-builder-canvas-area.pum-content > :nth-last-child(1 of :not(.pum-builder-canvas-title, .pum-builder-canvas-close-anchor, .pum-builder-canvas-close)) {
-				margin-bottom: 0;
-			}
-			.pum-builder-canvas-area.pum-scrollable > .pum-builder-canvas-title,
-			.pum-builder-canvas-area.pum-scrollable > .pum-builder-canvas-close {
-				translate: 0 var(--pum-builder-canvas-scroll-y, 0) !important;
-			}
-		`;
-		targetDocument.head.append( style );
-		ownedStyleElements.push( style );
 	};
 
 	const attachTitle = (): void => {
@@ -461,597 +313,20 @@ if ( ownedCanvas?.popup_id ) {
 		rootElement.style.setProperty( 'min-height', height, 'important' );
 	};
 
-	const fixedContainingBlock = (): HTMLElement | null => {
-		if ( ! canvas || ! canvasWindow ) {
-			return null;
-		}
-
-		let ancestor = canvas.parentElement;
-
-		while ( ancestor ) {
-			const style = canvasWindow.getComputedStyle( ancestor );
-			const contain = style.getPropertyValue( 'contain' );
-			const willChange = style.getPropertyValue( 'will-change' );
-			const createsContainingBlock =
-				[
-					'transform',
-					'translate',
-					'scale',
-					'rotate',
-					'perspective',
-					'filter',
-					'backdrop-filter',
-				].some( ( property ) => {
-					const value = style.getPropertyValue( property );
-
-					return '' !== value && 'none' !== value;
-				} ) ||
-				/(?:layout|paint|strict|content)/.test( contain ) ||
-				/(?:transform|perspective|filter)/.test( willChange ) ||
-				'auto' === style.getPropertyValue( 'content-visibility' );
-
-			if ( createsContainingBlock ) {
-				return ancestor;
-			}
-
-			ancestor = ancestor.parentElement;
-		}
-
-		return null;
-	};
-
-	type Matrix2D = {
-		a: number;
-		b: number;
-		c: number;
-		d: number;
-	};
-
-	const identityMatrix = (): Matrix2D => ( {
-		a: 1,
-		b: 0,
-		c: 0,
-		d: 1,
-	} );
-
-	const multiplyMatrices = (
-		left: Matrix2D,
-		right: Matrix2D
-	): Matrix2D => ( {
-		a: left.a * right.a + left.c * right.b,
-		b: left.b * right.a + left.d * right.b,
-		c: left.a * right.c + left.c * right.d,
-		d: left.b * right.c + left.d * right.d,
-	} );
-
-	const transformMatrix = ( transform: string ): Matrix2D | null => {
-		const match = transform.match( /^(matrix|matrix3d)\((.+)\)$/ );
-
-		if ( ! match ) {
-			return null;
-		}
-
-		const values = match[ 2 ].split( ',' ).map( Number );
-
-		if ( values.some( ( value ) => Number.isNaN( value ) ) ) {
-			return null;
-		}
-
-		if ( 'matrix' === match[ 1 ] && 6 === values.length ) {
-			const [ a, b, c, d ] = values;
-
-			return { a, b, c, d };
-		}
-
-		if ( 'matrix3d' === match[ 1 ] && 16 === values.length ) {
-			const isPlanar = [ 2, 3, 6, 7, 8, 9, 11, 14 ].every(
-				( index ) => 0 === values[ index ]
-			);
-
-			if ( ! isPlanar || 1 !== values[ 10 ] || 1 !== values[ 15 ] ) {
-				return null;
-			}
-
-			return {
-				a: values[ 0 ],
-				b: values[ 1 ],
-				c: values[ 4 ],
-				d: values[ 5 ],
-			};
-		}
-
-		return null;
-	};
-
-	const scaleValue = ( value: string ): number | null => {
-		const scale = parseFloat( value );
-
-		if ( Number.isNaN( scale ) ) {
-			return null;
-		}
-
-		return value.endsWith( '%' ) ? scale / 100 : scale;
-	};
-
-	const scaleMatrix = ( scale: string ): Matrix2D | null => {
-		if ( ! scale || 'none' === scale ) {
-			return identityMatrix();
-		}
-
-		const values = scale.trim().split( /\s+/ ).map( scaleValue );
-
-		if (
-			values.some( ( value ) => null === value ) ||
-			values.length < 1 ||
-			values.length > 3 ||
-			( 3 === values.length && 1 !== values[ 2 ] )
-		) {
-			return null;
-		}
-
-		return {
-			...identityMatrix(),
-			a: values[ 0 ] as number,
-			d: ( values[ 1 ] ?? values[ 0 ] ) as number,
-		};
-	};
-
-	const angleRadians = ( value: string ): number | null => {
-		const angle = parseFloat( value );
-
-		if ( Number.isNaN( angle ) ) {
-			return null;
-		}
-
-		if ( value.endsWith( 'deg' ) ) {
-			return ( angle * Math.PI ) / 180;
-		}
-
-		if ( value.endsWith( 'grad' ) ) {
-			return ( angle * Math.PI ) / 200;
-		}
-
-		if ( value.endsWith( 'turn' ) ) {
-			return angle * Math.PI * 2;
-		}
-
-		return value.endsWith( 'rad' ) || 0 === angle ? angle : null;
-	};
-
-	const rotateMatrix = ( rotate: string ): Matrix2D | null => {
-		if ( ! rotate || 'none' === rotate ) {
-			return identityMatrix();
-		}
-
-		const values = rotate.trim().split( /\s+/ );
-		let direction = 1;
-		let angleValue = '';
-
-		if ( 1 === values.length ) {
-			[ angleValue ] = values;
-		} else if ( 2 === values.length && 'z' === values[ 0 ] ) {
-			angleValue = values[ 1 ];
-		} else if ( 4 === values.length ) {
-			const axes = values.slice( 0, 3 ).map( Number );
-
-			if (
-				axes.some( ( axis ) => Number.isNaN( axis ) ) ||
-				0 !== axes[ 0 ] ||
-				0 !== axes[ 1 ] ||
-				0 === axes[ 2 ]
-			) {
-				return null;
-			}
-
-			direction = Math.sign( axes[ 2 ] );
-			angleValue = values[ 3 ];
-		} else {
-			return null;
-		}
-
-		const angle = angleRadians( angleValue );
-
-		if ( null === angle ) {
-			return null;
-		}
-
-		const cosine = Math.cos( angle * direction );
-		const sine = Math.sin( angle * direction );
-
-		return {
-			...identityMatrix(),
-			a: cosine,
-			b: sine,
-			c: -sine,
-			d: cosine,
-		};
-	};
-
-	const elementTransformMatrix = (
-		style: CSSStyleDeclaration
-	): Matrix2D | null => {
-		const transform = style.getPropertyValue( 'transform' );
-		const transformValue =
-			! transform || 'none' === transform
-				? identityMatrix()
-				: transformMatrix( transform );
-		const scale = scaleMatrix( style.getPropertyValue( 'scale' ) );
-		const rotate = rotateMatrix( style.getPropertyValue( 'rotate' ) );
-
-		if ( ! transformValue || ! scale || ! rotate ) {
-			return null;
-		}
-
-		// Individual transforms are applied translate, rotate, scale, then
-		// the transform list. Translation and origins affect only the affine
-		// offset, which is recovered from the element's viewport bounds.
-		return multiplyMatrices(
-			rotate,
-			multiplyMatrices( scale, transformValue )
-		);
-	};
-
-	const transformedAncestorMatrix = (
-		element: HTMLElement
-	): Matrix2D | null => {
-		if ( ! canvasWindow ) {
-			return null;
-		}
-
-		let current: HTMLElement | null = element;
-		let transformed = false;
-		const matrices: Matrix2D[] = [];
-
-		while ( current ) {
-			const style = canvasWindow.getComputedStyle( current );
-			const currentMatrix = elementTransformMatrix( style );
-
-			if ( ! currentMatrix ) {
-				return null;
-			}
-
-			if (
-				[ 'transform', 'translate', 'scale', 'rotate' ].some(
-					( property ) => {
-						const value = style.getPropertyValue( property );
-
-						return Boolean(
-							value && 'none' !== value && 'normal' !== value
-						);
-					}
-				)
-			) {
-				transformed = true;
-			}
-
-			matrices.push( currentMatrix );
-			current = current.parentElement;
-		}
-
-		return transformed
-			? matrices.reduce(
-					( matrix, currentMatrix ) =>
-						multiplyMatrices( currentMatrix, matrix ),
-					identityMatrix()
-			  )
-			: null;
-	};
-
-	const transformedContainingBlockPosition = (
-		element: HTMLElement,
-		viewportTop: number,
-		viewportLeft: number
-	): { left: number; top: number } | null => {
-		if ( ! canvasWindow ) {
-			return null;
-		}
-
-		const matrix = transformedAncestorMatrix( element );
-		const width = element.offsetWidth;
-		const height = element.offsetHeight;
-
-		if ( ! matrix || ! width || ! height ) {
-			return null;
-		}
-
-		const determinant = matrix.a * matrix.d - matrix.b * matrix.c;
-
-		if ( ! determinant ) {
-			return null;
-		}
-
-		const transformPoint = ( left: number, top: number ) => ( {
-			left: matrix.a * left + matrix.c * top,
-			top: matrix.b * left + matrix.d * top,
-		} );
-		const corners = [
-			transformPoint( 0, 0 ),
-			transformPoint( width, 0 ),
-			transformPoint( 0, height ),
-			transformPoint( width, height ),
-		];
-		const minLeft = Math.min( ...corners.map( ( point ) => point.left ) );
-		const minTop = Math.min( ...corners.map( ( point ) => point.top ) );
-		const bounds = element.getBoundingClientRect();
-		const relativeLeft = viewportLeft - ( bounds.left - minLeft );
-		const relativeTop = viewportTop - ( bounds.top - minTop );
-
-		return {
-			left:
-				( matrix.d * relativeLeft - matrix.c * relativeTop ) /
-				determinant,
-			top:
-				( -matrix.b * relativeLeft + matrix.a * relativeTop ) /
-				determinant,
-		};
-	};
-
-	const viewportToPosition = (
-		viewportTop: number,
-		viewportLeft: number,
-		position: string
-	): { left: number; top: number } => {
-		if ( ! canvas || ! canvasWindow ) {
-			return { left: viewportLeft, top: viewportTop };
-		}
-
-		const offsetParent = (
-			'fixed' === position
-				? fixedContainingBlock() || canvas.offsetParent
-				: canvas.offsetParent
-		) as HTMLElement | null;
-
-		if ( 'fixed' === position && ! offsetParent ) {
-			return { left: viewportLeft, top: viewportTop };
-		}
-
-		if ( offsetParent ) {
-			const transformedPosition = transformedContainingBlockPosition(
-				offsetParent,
-				viewportTop,
-				viewportLeft
-			);
-
-			if ( transformedPosition ) {
-				return {
-					left:
-						transformedPosition.left +
-						offsetParent.scrollLeft -
-						offsetParent.clientLeft,
-					top:
-						transformedPosition.top +
-						offsetParent.scrollTop -
-						offsetParent.clientTop,
-				};
-			}
-
-			const parentBounds = offsetParent.getBoundingClientRect();
-
-			return {
-				left:
-					viewportLeft -
-					parentBounds.left +
-					offsetParent.scrollLeft -
-					offsetParent.clientLeft,
-				top:
-					viewportTop -
-					parentBounds.top +
-					offsetParent.scrollTop -
-					offsetParent.clientTop,
-			};
-		}
-
-		return {
-			left: viewportLeft + canvasWindow.scrollX,
-			top: viewportTop + canvasWindow.scrollY,
-		};
-	};
-
-	const getCanvasTransformSignature = (
-		element: HTMLElement,
-		targetWindow: Window
-	): string => {
-		const signatures: string[] = [];
-		let current: HTMLElement | null = element;
-
-		while ( current ) {
-			const style = targetWindow.getComputedStyle( current );
-
-			signatures.push(
-				[
-					'transform',
-					'transform-origin',
-					'translate',
-					'scale',
-					'rotate',
-				]
-					.map( ( property ) => style.getPropertyValue( property ) )
-					.join( '|' )
-			);
-			current = current.parentElement;
-		}
-
-		return signatures.join( '||' );
-	};
-
-	const hasCanvasTransform = ( style: CSSStyleDeclaration ): boolean =>
-		[ 'transform', 'translate', 'scale', 'rotate' ].some( ( property ) => {
-			const value = style.getPropertyValue( property );
-
-			return Boolean( value && 'none' !== value && 'normal' !== value );
-		} );
-
-	function mirrorPopupGeometry(): void {
+	const mirrorPopupGeometry = (): void => {
 		if ( ! canvas || ! canvasWindow ) {
 			return;
 		}
 
-		const isCustom = 'custom' === display.size;
-		const isResponsive = ! isCustom && 'auto' !== display.size;
-		const height =
-			isCustom && ! isEnabled( display.custom_height_auto )
-				? viewportRelativeLength( display.custom_height, 'vh' )
-				: 'auto';
-		let vertical = 'center';
-		let horizontal = 'center';
-		let width = 'auto';
-
-		if ( isCustom ) {
-			width = viewportRelativeLength( display.custom_width );
-		} else if ( isResponsive ) {
-			width = responsiveWidth( display.size, canvasWindow.innerWidth );
-		}
-
-		( display.location || 'center' ).split( ' ' ).forEach( ( part ) => {
-			if ( 'top' === part || 'bottom' === part ) {
-				vertical = part;
-			} else if ( 'left' === part || 'right' === part ) {
-				horizontal = part;
-			}
-		} );
-
-		const requestedPosition =
-			isEnabled( display.position_fixed ) &&
-			! ( isResponsive && canvasWindow.innerWidth <= 1024 )
-				? 'fixed'
-				: 'absolute';
-		const properties: Record< string, string > = {
-			position: requestedPosition,
-			width,
-			height,
-			'min-width': isResponsive
-				? `min(${ viewportRelativeLength(
-						display.responsive_min_width
-				  ) }, calc(100vw - 20px))`
-				: '0',
-			'max-width': isResponsive
-				? `min(${ viewportRelativeLength(
-						display.responsive_max_width
-				  ) }, calc(100vw - 20px))`
-				: 'calc(100vw - 20px)',
-			'overflow-y': isEnabled( display.scrollable ) ? 'auto' : 'visible',
-			margin: '0',
-			top: '0',
-			bottom: 'auto',
-			left: '0',
-			right: 'auto',
-		};
-
-		Object.entries( properties ).forEach( ( [ property, value ] ) => {
-			if ( value ) {
-				setCanvasStyle( property, value, 'important' );
-			}
-		} );
-
-		const bounds = canvas.getBoundingClientRect();
-		const closeButton = canvas.querySelector< HTMLElement >(
-			':scope > .pum-builder-canvas-close'
-		);
-		const closeBounds = closeButton?.getBoundingClientRect();
-		let visualTop = 0;
-		let visualRight = bounds.width;
-		let visualBottom = bounds.height;
-		let visualLeft = 0;
-
-		if ( closeBounds && ( closeBounds.width || closeBounds.height ) ) {
-			visualTop = Math.min( visualTop, closeBounds.top - bounds.top );
-			visualRight = Math.max(
-				visualRight,
-				closeBounds.right - bounds.left
-			);
-			visualBottom = Math.max(
-				visualBottom,
-				closeBounds.bottom - bounds.top
-			);
-			visualLeft = Math.min( visualLeft, closeBounds.left - bounds.left );
-		}
-
-		let viewportTop = ( canvasWindow.innerHeight - bounds.height ) / 2;
-		let viewportLeft = ( canvasWindow.innerWidth - bounds.width ) / 2;
-
-		if ( 'top' === vertical ) {
-			viewportTop = numericOffset( display.position_top );
-		} else if ( 'bottom' === vertical ) {
-			viewportTop =
-				canvasWindow.innerHeight -
-				bounds.height -
-				numericOffset( display.position_bottom );
-		}
-
-		if ( 'left' === horizontal ) {
-			viewportLeft = numericOffset( display.position_left );
-		} else if ( 'right' === horizontal ) {
-			viewportLeft =
-				canvasWindow.innerWidth -
-				bounds.width -
-				numericOffset( display.position_right );
-		}
-
-		const edgeGap = 10;
-		const topEdgeGap = canvas.ownerDocument.body.classList.contains(
-			'admin-bar'
-		)
-			? 42
-			: edgeGap;
-		const minTop = topEdgeGap - visualTop;
-		const maxTop = canvasWindow.innerHeight - edgeGap - visualBottom;
-		const minLeft = edgeGap - visualLeft;
-		const maxLeft = canvasWindow.innerWidth - edgeGap - visualRight;
-		const isOversized =
-			visualBottom - visualTop + topEdgeGap + edgeGap >
-			canvasWindow.innerHeight;
-
-		viewportTop = Math.max( minTop, Math.min( viewportTop, maxTop ) );
-		viewportLeft = Math.max( minLeft, Math.min( viewportLeft, maxLeft ) );
-
-		const position = isOversized ? 'absolute' : requestedPosition;
-		setCanvasStyle( 'position', position, 'important' );
-		const coordinates = viewportToPosition(
-			viewportTop,
-			viewportLeft,
-			position
-		);
-		const canvasStyle = canvasWindow.getComputedStyle( canvas );
-		canvasTransformSignature = getCanvasTransformSignature(
+		canvasTransformSignature = updatePopupGeometry( {
 			canvas,
-			canvasWindow
-		);
-
-		if ( hasCanvasTransform( canvasStyle ) ) {
-			const positionedBounds = canvas.getBoundingClientRect();
-			const transformOffset = viewportToPosition(
-				positionedBounds.top,
-				positionedBounds.left,
-				position
-			);
-
-			coordinates.top -= transformOffset.top;
-			coordinates.left -= transformOffset.left;
-		}
-
-		setCanvasStyle(
-			'top',
-			`${ Math.round( coordinates.top * 100 ) / 100 }px`,
-			'important'
-		);
-		setCanvasStyle( 'bottom', 'auto', 'important' );
-		setCanvasStyle(
-			'left',
-			`${ Math.round( coordinates.left * 100 ) / 100 }px`,
-			'important'
-		);
-		setCanvasStyle( 'right', 'auto', 'important' );
-
-		if ( isOversized ) {
-			setRootMinHeight(
-				`${ Math.ceil(
-					canvasWindow.scrollY + viewportTop + visualBottom + edgeGap
-				) }px`
-			);
-		} else {
-			restoreRootMinHeight();
-		}
-	}
+			display,
+			restoreRootMinHeight,
+			setCanvasStyle,
+			setRootMinHeight,
+			targetWindow: canvasWindow,
+		} );
+	};
 
 	const scheduleGeometry = (): void => {
 		if ( null !== geometryFrame || ! canvasWindow ) {
@@ -1176,28 +451,37 @@ if ( ownedCanvas?.popup_id ) {
 			canvasWindow = targetWindow;
 			rootElement = targetDocument.documentElement;
 
-			if ( iframe ) {
-				if ( 'MutationObserver' in window ) {
-					targetObserver = new window.MutationObserver(
-						scheduleAdoptionFromMutations
-					);
-					targetObserver.observe( targetDocument.documentElement, {
-						attributes: true,
-						attributeFilter: selectorAttributeFilter(
-							display.canvas_selector
-						),
-						childList: true,
-						subtree: true,
-					} );
-				}
+			if ( iframe && 'MutationObserver' in window ) {
+				targetObserver = new window.MutationObserver(
+					scheduleAdoptionFromMutations
+				);
+				targetObserver.observe( targetDocument.documentElement, {
+					attributes: true,
+					attributeFilter: selectorAttributeFilter(
+						display.canvas_selector
+					),
+					childList: true,
+					subtree: true,
+				} );
 			}
 
 			targetWindow.addEventListener( 'resize', scheduleGeometry );
 		}
 
 		if ( iframe ) {
-			copyPopupStyles( targetDocument );
-			attachCanvasStyles( targetDocument );
+			ownedStyleElements.push(
+				...copyPopupStyles(
+					document,
+					targetDocument,
+					display.style_selectors || [],
+					adoptCanvas
+				)
+			);
+			const attachedStyles = attachCanvasStyles( targetDocument );
+
+			if ( attachedStyles ) {
+				ownedStyleElements.push( attachedStyles );
+			}
 		}
 
 		let targetCanvas = targetDocument.querySelector< HTMLElement >(
